requirements.txt
annotated-types==0.7.0
audioread==3.0.1
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
docstring_parser==0.16
Flask==3.0.3
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.0
google-cloud-translate==3.16.0
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
googletrans==3.0.0
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.62.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
langdetect==1.0.9
lazy_loader==0.4
libretranslatepy==2.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
MarkupSafe==2.1.5
msgpack==1.0.8
numba==0.60.0
numpy==2.0.1
packaging==24.1
platformdirs==4.2.2
pooch==1.8.2
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.5.1
scipy==1.13.1
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.4.0
threadpoolctl==3.5.0
tqdm==4.66.5
translate==3.6.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
